requirements.txt
django
djangorestframework
markdown
django-filter
coreapi
Pygments
tzlocal
django-cors-headers